refactor: 코드 리팩토링 및 로직확인용 콘솔 추가
